Advent of Code 2017 Solutions by @bericp1 v0.1.0 AdventOfCode2017

A solution set for the advent of code 2017.

Functions

loop()

loop() :: integer()

Main loop. Ask the user what they want to do, process the selection, and process its result.

Will accept the following results:

:bail will bail out of the loop and exit with a zero (successful) status code

:continue will restart the loop

main(args)

Called as the entry point by escript.

process_result(result)

process_result(tuple() | atom()) :: atom()

Process a result from process_selection/1 for use in loop/0.

Will produce side effects.

result should be one of the following:

{:ok, message}: run completed successfully and has a message to share with the user. Loop should continue.

{:error, message}: run failed for some reason and has a message about the error to share with the user. Loop should continue.

Anything else (e.g. :bail or :continue) will be returned directly.

process_selection(selection)

process_selection(charlist()) :: atom() | tuple()

Process a selection received from select/0.

selection should be one of:

a valid number matching an implemented advent “day” (a number between 1 and 1)

one of qquitexit

run(day)

run(integer()) :: tuple()

Run a day’s puzzle by number and executing its run method.

select()

select() :: charlist()

Read in a selection from stdin.