Advent of Code 2017 Solutions by @bericp1 v0.1.0 AdventOfCode2017.Day1

Day 1: Inverse Captcha

Functions

at_wraparound(list, index)

at_wraparound(list :: [...], index :: integer()) :: any()

Enum.at/2 but will wrap around if index is greater than the length/1 of list.

byte_to_int(byte)

byte_to_int(char()) :: integer()

Convert a utf8 byte representation of an integer into the integer it represents.

compare_digits(digit_a, digit_b)

compare_digits(digit_a :: integer(), digit_b :: integer()) :: integer()

Compares two digits and if they are equal, returns their value. Otherwise returns 0.

See also compare_digits/3

compare_digits(digits, digit_a, index_b)

compare_digits(
  list :: [integer(), ...],
  digit_a :: integer(),
  index_b :: integer()
) :: integer()

Given a list, will compare digit_a with the digit at index_b, returning their value if equivleant and 0 otherwise.

See compare_digits/2, at_wraparound/2

run()

Run the Day 1 puzzle to gather input from the user.

str_to_int_list(str)

str_to_int_list(str :: charlist()) :: [integer(), ...]

Convert a string (charlist) to a list of integers.