Advent of Code 2017 Solutions by @bericp1 v0.1.0 AdventOfCode2017

A solution set for the advent of code 2017.

Functions

loop()

Main loop. Ask the user what they want to do, process the selection, and process its result.

Will accept the following results:

:bail will bail out of the loop and exit with a zero (successful) status code

:continue will restart the loop

main(args)

Called as the entry point by escript.

process_result(fwd)

Forward all other selection results straight through without transforming them.

process_selection(selection)

Handle number selection by running and returning the result of the corrresponding day

run()

Error case when no day is provided.

run(day)

Run a day’s puzzle by number and executing its run method.

select()

Read in a selection from stdin