Advent of Code 2017 Solutions by @bericp1 v0.1.0 AdventOfCode2017.Day1

Day 1: Inverse Captcha

Functions

byte_to_int(byte)

run()

solve(digits)

solve(digits :: [integer(), ...]) :: integer()

Circularly find the sum of all digits that match the next digit in the list

solve(arg1, sum)

solve_str(str)

solve_str(str :: charlist()) :: integer()

Circularly find the sum of all digits that match the next digit in the string of digits

str_to_int_list(str)

str_to_int_list(str :: charlist()) :: [integer(), ...]